requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
pandas>=2.2.2